Create New class feature
Validate on the backend if the class and student exists in the db or not, if does not exist then create. Update corresponding tables accordingly, notably the enrollment table as per the ER diagram. The enrollment table has unique instructor_id, course_id set as unique values so that an instructor can have a 2 courses with the same name.   If class exists, use the existing class_id.

Might need to inform the user in case the class already exists on the front-end.
Added a toast to notify the user that the class has been created and then the user is navigated to the class-management tab.