Added Custom marking scheme
Modified ManuelExamKey and ConfirmExamKey to add a custom marking scheme that is stored in the backend and passed to evaluation.
I added confirmation that input has been entered in New Exam and the custom marking scheme pop up window.
I readded  react-app-rewired-start to not have to restart the frontend container each time