Custom Bubble Sheet with a single page
Modified logic in ReviewExams to pass numQuestion and examtype to studentScores and if numQuestions <=100 and examType is custom don't run preprocessingCSV as we do not need to combinecsv for 2 seperate pages, you only need to get Results.csv.
In the Backend we are now checking in studentScores if numQuestions <=100 and examType is custom then get the result from Results.csv instead of combined.csv
UploadExam also makes this check to see if it needs to create 2 folder for each page. The same is done in saveExamKey, copyTemplate and for GenerateEvaluation.