The report feature is fullly functional
Instructors can reply edit to the students message, edit the grade and close reports, they can reopen reports, editing is disabled after closing a report it has to be reopened to be edited again, students cannot resubmit a pending report, students cannot submit a blank report.
Images is displayed on the report.
Student reports and instructor replies are saved in the db.